+++ packages/houdini/src/lib/config.ts
@@ -59,6 +59,7 @@ export class Config {
 	routesDir: string
 	schemaPollInterval: number | null
 	schemaPollTimeout: number
+	schemaPollWriteToDisk: boolean = false
 	schemaPollHeaders:
 		| ((env: any) => Record<string, string>)
 		| Record<string, string | ((env: any) => string)>
@@ -163,6 +164,7 @@ export class Config {
 		this.routesDir = path.join(this.projectRoot, 'src', 'routes')
 		this.schemaPollInterval = watchSchema?.interval === undefined ? 2000 : watchSchema.interval
 		this.schemaPollTimeout = watchSchema?.timeout ?? 30000
+		this.schemaPollWriteToDisk = watchSchema?.writePolledSchema ?? true
 		this.schemaPollHeaders = watchSchema?.headers ?? {}
 		this.rootDir = path.join(this.projectRoot, this.runtimeDir)
 		this.persistedQueriesPath =
@@ -211,17 +213,24 @@ export class Config {
 		// documents
 		for (const plugin of this.plugins) {
 			const runtimeDir = this.pluginRuntimeSource(plugin)
+			const staticDir = this.pluginStaticRuntimeSource(plugin)
 
 			// skip plugins that dont' include runtimes
-			if (!runtimeDir) {
+			if (!runtimeDir && !staticDir) {
 				continue
 			}
 
-			// the include path is relative to root of the vite project
-			const includePath = path.relative(this.projectRoot, runtimeDir)
+			for (const dir of [runtimeDir, staticDir]) {
+				if (!dir) {
+					continue
+				}
 
-			// add the plugin's directory to the include pile
-			include.push(`${includePath}/**/*{${extensions.join(',')}}`)
+				// the include path is relative to root of the vite project
+				const includePath = path.relative(this.projectRoot, dir)
+
+				// add the plugin's directory to the include pile
+				include.push(`${includePath}/**/*{${extensions.join(',')}}`)
+			}
 		}
 
 		return include
@@ -300,6 +309,19 @@ export class Config {
 			typeof plugin.includeRuntime === 'string'
 				? plugin.includeRuntime
 				: plugin.includeRuntime?.[this.module]
+		)
+	}
+
+	pluginStaticRuntimeSource(plugin: PluginMeta) {
+		if (!plugin.staticRuntime) {
+			return null
+		}
+
+		return path.join(
+			path.dirname(plugin.filepath),
+			typeof plugin.staticRuntime === 'string'
+				? plugin.staticRuntime
+				: plugin.staticRuntime?.[this.module]
 		)
 	}
 
@@ -602,6 +624,10 @@ export class Config {
 
 	pluginRuntimeDirectory(name: string) {
 		return path.join(this.pluginDirectory(name), 'runtime')
+	}
+
+	pluginStaticRuntimeDirectory(name: string) {
+		return path.join(this.pluginDirectory(name), 'static')
 	}
 
 	get pluginRootDirectory() {
@@ -1157,11 +1183,14 @@ export async function getConfig({
 			// we might have to pull the schema first
 			if (apiURL) {
 				// make sure we don't have a pattern pointing to multiple files and a remove URL
-				if (fs.glob.hasMagic(_config.schemaPath)) {
+				if (fs.glob.hasMagic(_config.schemaPath) && _config.schemaPollWriteToDisk) {
 					console.log(
 						`⚠️  Your houdini configuration contains an apiUrl and a path pointing to multiple files.
-	This will prevent your schema from being pulled.`
+	This will prevent your schema from being written to disk. If this is expected, please set the writePolledSchema value to false.`
 					)
+
+					// Don't write the schema to disk, since it'll error out
+					_config.schemaPollWriteToDisk = false
 				}
 				// we might have to create the file
 				else if (!(await fs.readFile(_config.schemaPath))) {
@@ -1170,7 +1199,9 @@ export async function getConfig({
 						(await pullSchema(
 							apiURL,
 							_config.schemaPollTimeout,
-							_config.schemaPath
+							_config.schemaPath,
+							{},
+							_config.schemaPollWriteToDisk
 						)) !== null
 				}
 			}