+++ packages/adapter-static/CHANGELOG.md
@@ -1,5 +1,26 @@
 # houdini-adapter-static
 
+## 1.5.5-next.3
+
+### Patch Changes
+
+-   Updated dependencies [[`20bbeeda`](https://github.com/HoudiniGraphql/houdini/commit/20bbeedae7e586191dcc0d51685017de659df5bd)]:
+    -   houdini@2.0.0-next.3
+
+## 1.5.5-next.2
+
+### Patch Changes
+
+-   Updated dependencies [[`23f56f35`](https://github.com/HoudiniGraphql/houdini/commit/23f56f35f8fadcd5ee45e500460b6387a68dd6df)]:
+    -   houdini@2.0.0-next.2
+
+## 1.5.5-next.1
+
+### Patch Changes
+
+-   Updated dependencies [[`e92beaf4`](https://github.com/HoudiniGraphql/houdini/commit/e92beaf4a680082a644726abfb719cfeda72ab8e), [`83db8ab2`](https://github.com/HoudiniGraphql/houdini/commit/83db8ab2678d5e86c325890f6f3608f16ea3c81f)]:
+    -   houdini@2.0.0-next.1
+
 ## 1.5.5-next.0
 
 ### Patch Changes